+++ gym_super_mario_bros/__init__.py
@@ -2,7 +2,6 @@
 from .smb_env import SuperMarioBrosEnv
 from .smb2_env import SuperMarioBros2Env
 from .smb3_env import SuperMarioBros3Env
-from .smb_random_stages_env import SuperMarioBrosRandomStagesEnv
 from ._registration import make
 
 
@@ -12,5 +11,4 @@ __all__ = [
     'SuperMarioBrosEnv',
     'SuperMarioBros2Env',
     'SuperMarioBros3Env',
-    'SuperMarioBrosRandomStagesEnv',
 ]