+++ gym_super_mario_bros/_roms/__init__.py
@@ -1,6 +1,7 @@
 """Methods for ROM file management."""
 from .decode_target import decode_target
-from .rom_path import rom_path
+from .rom_path import smb1_rom_path
+from .rom_path import smb2jp_rom_path
 from .rom_path import smb2_rom_path
 from .rom_path import smb3_rom_path
 
@@ -8,7 +9,8 @@ from .rom_path import smb3_rom_path
 # explicitly define the outward facing API of this package
 __all__ = [
     decode_target.__name__,
-    rom_path.__name__,
+    smb1_rom_path.__name__,
+    smb2jp_rom_path.__name__,
     smb2_rom_path.__name__,
     smb3_rom_path.__name__,
 ]